Added some functions + tests to backend
Added getAllPosts function and clearDatabase function to functions.js. Added connectability from express server (app.js) to the getAllPosts function in functions.js. We will not make a connection to clearDatabase function because that's a pretty major security flaw.

Also added tests for the express server and the functions.js script for getAllPosts.